+++ CHANGELOG.md
@@ -10,6 +10,7 @@ non-deterministic and rely on pulling data from the network.
 - Fixed a bug where `quilt inspect` would panic when given a relative path.
 - Use the latest release of OVS (2.7.2).
 - Remove support for invariants.
+- Remove support for placement based on service groups.
 
 Release 0.2.0
 -------------